Extended svg command with option to generate bottom.
When option -width and -height are used for svg command, than generates optimal sizes such that the puzzle and the bottom fit within given sizes. It assumes that the width is larger than the height. The -margin option can be used to specify a margin. Also works when -width and -height are not used.