Fixing the -bottom command without -height and -width
Steven Parker (perlygatekeeper) reported a problem if the -bottom
option is used with the svg subcommand. This must have gone broke
when I added the -height and -width options and forgot to test the
program without these options. He did suggest a fix to the program.
I found a slightly beter solution, which also streamlines the code.

I also removed and commented some debugging code.