Adding -space option for svg subcommand
I added the option -space, which allows you to specify the space between the pieces. Most DIY-laser cutters make a very narrow cut, which means that if you make a puzzle, the pieces will fit very tightly together. To avoid this, the idea is to make two cuts and throw away the material in between. For this, I had to write a new algorithm. The old algorithm (still used when no value or a non-positive value for -space is used) would generate lines from top to down. The new algorithm generate a cut for each piece. For this reason the algorithm needs to track the sides of the pieces. This method can also be used to cut the border by following the (in)side of the border. Because now these are cut one-by-one, all the corners will be round, and the shape of the pieces and border is a little different from the single line cut. I am thinking about adding the option -round to implement this same shape in the single cut algorithm.